1. Rename anofora_funct files 2. Add descriptions at README file
